Treat all ad-hoc downloads as dependencies
For each ad-hoc download, we now create a single-dependency
configuration and declare an exclusive-content Ivy repository to resolve
that dependency.  This moves downloads out of the local build tree and
instead retains them in the user's Gradle cache under `~/.gradle`.

I like the idea of caching downloads outside of any single WALA tree,
since the downloads themselves don't change from build to build.
However, this approach also has two disadvantages:

1. Gradle downloads all dependencies that the task graph will need
   before starting the tasks themselves:  even the very first task
   must wait until the last dependency is downloaded.  Additionally,
   while downloads can be concurrent, I see less concurrency that we
   achieve when evaluating the task graph.  In practice, many of the
   largest downloads are happening with no other concurrent activity,
   thereby leaving a lot of potential parallelism on the table.

2. The flexibility of Ivy repository configuration means that you can
   effectively download from any URL.  But Gradle has some strongly
   ingrained assumptions that each dependency has a name, a version,
   and an extension.  Most of our ad-hoc downloads can be treated in
   that way, but things get awkward for the few that cannot.

This was an interesting experiment, and worth preserving for future
reference.  But I don't plan to propose it for inclusion in the main
WALA branch at this time.